chore: renaming balancedbinarytree -> avltree
