+++ src/test/java/com/github/avl_tree/TestHeight.java
@@ -1,4 +1,4 @@
-package com.github.balanced_binary_tree;
+package com.github.avl_tree;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -14,13 +14,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 public class TestHeight {
     @Test
     public void testHeightEmptyTreeShouldBeMinusOne() {
-        BalancedBinarySearchTree<Integer> avl = new BalancedBinarySearchTree<>();
+        AvlTree<Integer> avl = new AvlTree<>();
         assertEquals(avl.height(), -1);
     }
 
     @Test
     public void testHeightLinkedListLikeTree() {
-        BalancedBinarySearchTree<Integer> avl = new BalancedBinarySearchTree<>();
+        AvlTree<Integer> avl = new AvlTree<>();
         for (int i = -7; i < 0; i++) {
             avl.add(i);
         }
@@ -34,7 +34,7 @@ public class TestHeight {
         for (int i = 0; i < 1000; i++) {
             list.add(i);
         }
-        BalancedBinarySearchTree<Integer> avl = new BalancedBinarySearchTree<>();
+        AvlTree<Integer> avl = new AvlTree<>();
         Random random = new Random(seed);
         for (int i = 0; i < 1000; i++) {
             int numberToAdd = list.get(Math.abs(random.nextInt() % list.size()));
@@ -48,7 +48,7 @@ public class TestHeight {
     @ParameterizedTest
     @ValueSource(ints = { 1, 2, 3, 4 })
     public void testSequentialAddHeightShouldBeInInterval(int seed) {
-        BalancedBinarySearchTree<Integer> avl = new BalancedBinarySearchTree<>();
+        AvlTree<Integer> avl = new AvlTree<>();
         for (int i = 0; i < 1000; i++) {
             avl.add(i);
         }
@@ -59,7 +59,7 @@ public class TestHeight {
 
     @Test
     public void testTreeShouldRotateLeftOnce() {
-        BalancedBinarySearchTree<Integer> avl = new BalancedBinarySearchTree<>();
+        AvlTree<Integer> avl = new AvlTree<>();
         avl.add(1);
         assertEquals(avl.height(), 0);
         avl.add(2);
@@ -70,7 +70,7 @@ public class TestHeight {
 
     @Test
     public void testTreeShouldRotateRightLeftOnce() {
-        BalancedBinarySearchTree<Integer> avl = new BalancedBinarySearchTree<>();
+        AvlTree<Integer> avl = new AvlTree<>();
         avl.add(0);
         assertEquals(avl.height(), 0);
         avl.add(-3);
@@ -85,7 +85,7 @@ public class TestHeight {
 
     @Test
     public void testTreeShouldRotateRightOnce() {
-        BalancedBinarySearchTree<Integer> avl = new BalancedBinarySearchTree<>();
+        AvlTree<Integer> avl = new AvlTree<>();
         avl.add(-1);
         assertEquals(avl.height(), 0);
         avl.add(-2);
@@ -96,7 +96,7 @@ public class TestHeight {
 
     @Test
     public void testTreeShouldRotateLeftRightOnce() {
-        BalancedBinarySearchTree<Integer> avl = new BalancedBinarySearchTree<>();
+        AvlTree<Integer> avl = new AvlTree<>();
         avl.add(0);
         assertEquals(avl.height(), 0);
         avl.add(3);
